requirements.txt
xlsxwriter
jsonschema
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
openpyxl
pytest
pytest-asyncio
